+++ protoc_plugin/test/goldens/messageGenerator.pb.dart
@@ -3,43 +3,20 @@ class PhoneNumber extends $pb.GeneratedMessage {
 
   PhoneNumber._();
 
-  factory PhoneNumber.fromBuffer(
-    $core.List<$core.int> data, [
-    $pb.ExtensionRegistry registry = $pb.ExtensionRegistry.EMPTY,
-  ]) => create()..mergeFromBuffer(data, registry);
-  factory PhoneNumber.fromJson(
-    $core.String json, [
-    $pb.ExtensionRegistry registry = $pb.ExtensionRegistry.EMPTY,
-  ]) => create()..mergeFromJson(json, registry);
+  factory PhoneNumber.fromBuffer($core.List<$core.int> data, [$pb.ExtensionRegistry registry = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(data, registry);
+  factory PhoneNumber.fromJson($core.String json, [$pb.ExtensionRegistry registry = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(json, registry);
 
-  static final $pb.BuilderInfo _i =
-      $pb.BuilderInfo(
-          _omitMessageNames ? '' : 'PhoneNumber',
-          createEmptyInstance: create,
-        )
-        ..aQS(1, _omitFieldNames ? '' : 'number')
-        ..e<PhoneNumber_PhoneType>(
-          2,
-          _omitFieldNames ? '' : 'type',
-          $pb.PbFieldType.OE,
-          defaultOrMaker: PhoneNumber_PhoneType.MOBILE,
-          valueOf: PhoneNumber_PhoneType.valueOf,
-          enumValues: PhoneNumber_PhoneType.values,
-        )
-        ..a<$core.String>(
-          3,
-          _omitFieldNames ? '' : 'name',
-          $pb.PbFieldType.OS,
-          defaultOrMaker: '\$',
-        )
-        ..aOS(4, _omitFieldNames ? '' : 'deprecatedField');
+  static final $pb.BuilderInfo _i = $pb.BuilderInfo(_omitMessageNames ? '' : 'PhoneNumber', createEmptyInstance: create)
+    ..aQS(1, _omitFieldNames ? '' : 'number')
+    ..e<PhoneNumber_PhoneType>(2, _omitFieldNames ? '' : 'type', $pb.PbFieldType.OE, defaultOrMaker: PhoneNumber_PhoneType.MOBILE, valueOf: PhoneNumber_PhoneType.valueOf, enumValues: PhoneNumber_PhoneType.values)
+    ..a<$core.String>(3, _omitFieldNames ? '' : 'name', $pb.PbFieldType.OS, defaultOrMaker: '\$')
+    ..aOS(4, _omitFieldNames ? '' : 'deprecatedField')
+  ;
 
   @$core.Deprecated('See https://github.com/google/protobuf.dart/issues/998.')
   PhoneNumber clone() => PhoneNumber()..mergeFromMessage(this);
   @$core.Deprecated('See https://github.com/google/protobuf.dart/issues/998.')
-  PhoneNumber copyWith(void Function(PhoneNumber) updates) =>
-      super.copyWith((message) => updates(message as PhoneNumber))
-          as PhoneNumber;
+  PhoneNumber copyWith(void Function(PhoneNumber) updates) => super.copyWith((message) => updates(message as PhoneNumber)) as PhoneNumber;
 
   @$core.override
   $pb.BuilderInfo get info_ => _i;
@@ -50,10 +27,7 @@ class PhoneNumber extends $pb.GeneratedMessage {
   PhoneNumber createEmptyInstance() => create();
   static $pb.PbList<PhoneNumber> createRepeated() => $pb.PbList<PhoneNumber>();
   @$core.pragma('dart2js:noInline')
-  static PhoneNumber getDefault() =>
-      _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<PhoneNumber>(
-        create,
-      );
+  static PhoneNumber getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<PhoneNumber>(create);
   static PhoneNumber? _defaultInstance;
 
   @$pb.TagNumber(1)
@@ -97,9 +71,6 @@ class PhoneNumber extends $pb.GeneratedMessage {
   void clearDeprecatedField() => $_clearField(4);
 }
 
-const $core.bool _omitFieldNames = $core.bool.fromEnvironment(
-  'protobuf.omit_field_names',
-);
-const $core.bool _omitMessageNames = $core.bool.fromEnvironment(
-  'protobuf.omit_message_names',
-);
+
+const $core.bool _omitFieldNames = $core.bool.fromEnvironment('protobuf.omit_field_names');
+const $core.bool _omitMessageNames = $core.bool.fromEnvironment('protobuf.omit_message_names');