+++ protoc_plugin/test/service_generator_test.dart
@@ -31,7 +31,10 @@ void main() {
 
     final serviceWriter = IndentingWriter();
     fg.serviceGenerators[0].generate(serviceWriter);
-    expectGolden(serviceWriter.toString(), 'serviceGenerator.pb.dart');
+    expectGolden(
+      serviceWriter.emitSource(format: true),
+      'serviceGenerator.pb.dart',
+    );
     expectGolden(fg.generateJsonFile(), 'serviceGenerator.pbjson.dart');
   });
 }